worker/requirements.txt
livekit
livekit-agents
faster-whisper
python-dotenv
numpy
